Adds a covariance matrix unit test

+++ Header/Kernels.h
@@ -14,7 +14,7 @@
 
 #include <stdint.h>
 
-__device__ __host__ unsigned int upperTriangularLength(unsigned int numRows);
+__device__ __host__ uint64_t upperTriangularLength(unsigned int numRows);
 
 __global__ void test();
 
@@ -32,7 +32,7 @@ __global__ void parallelMeanUnroll2(float* d_inputArray, uint64_t inputLength, f
 //---------------------
 __global__ void outerProductSmartBruteForce(float* resultMatrix, float* vec, int vectorLength);
 
-__global__ void outerProductSmartBruteForceLessThreads(float* resultMatrix, float* vec, int vectorLength);
+__global__ void outerProductSmartBruteForceLessThreads(float* resultMatrix, float* vec, uint64_t vectorLength);
 
 //Specialised outer product for DSPSR
 __global__ void outerProductUpperTri(cuFloatComplex* resultMatrix, cuFloatComplex* vec, unsigned int vectorLength);
